Copied in a bunch of mfc's
I don't know why it didn't catch the rest of these files last commit.
All I wanted to copy in was these from last commit.

in AFX.H
in AFXCONV.H
in AFXCTL.H
in AFXDAO.H
in AFXDB.H
in AFXDD_.H
in AFXDISP.H
in AFXDLL_.H
in AFXDLLX.H
in AFXEXT.H
in AFXINET.H
in AFXISAPI.H
in AFXOLE.H
in AFXPRIV.H
in AFXSOCK.H
in AFXSTAT_.H
in AFXTEMPL.H
in AFXTLS_.H
in AFXV_W32.H
in AFXWIN.H